pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "github-mcp"
version = "0.1.0"
description = "Model Context Protocol (MCP) server for GitHub integration"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Garot Conklin" }
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.4.2",
    "PyGithub>=2.1.1",
    "python-dotenv>=1.0.0",
    "sse-starlette>=1.6.5",
    "httpx>=0.25.0",
    "githubauthlib>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.10.1",
    "isort>=5.12.0",
    "mypy>=1.6.1",
    "ruff>=0.1.3",
]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"] 